+++ typescript/agentkit/src/wallet-providers/cdpWalletProvider.ts
@@ -142,7 +142,7 @@ export class CdpWalletProvider extends EvmWalletProvider {
     if (config.apiKeyName && config.apiKeyPrivateKey) {
       Coinbase.configure({
         apiKeyName: config.apiKeyName,
-        privateKey: config.apiKeyPrivateKey,
+        privateKey: config.apiKeyPrivateKey?.replace(/\\n/g, "\n"),
         source: "agentkit",
         sourceVersion: version,
       });