+++ typescript/agentkit/src/action-providers/index.ts
@@ -9,6 +9,7 @@ export * from "./cdp";
 export * from "./erc20";
 export * from "./erc721";
 export * from "./farcaster";
+export * from "./jupiter";
 export * from "./pyth";
 export * from "./moonwell";
 export * from "./morpho";