+++ typescript/agentkit/CHANGELOG.md
@@ -6,6 +6,9 @@
 
 - Add support for fauceting SOL on `solana-devnet`.
 
+### Fixed
+- Fixed handling of `CDP_API_KEY_PRIVATE_KEY` by moving parsing into CDP classes.
+
 ## [0.2.1] - 2025-02-18
 
 ### Added