+++ typescript/agentkit/CHANGELOG.md
@@ -10,7 +10,8 @@
 ### Fixed
 
 - Fixed handling of `CDP_API_KEY_PRIVATE_KEY` by moving parsing into CDP classes.
-- Fixed for `splActionProvider` `get_balance` `TokenAccountNotFoundError`
+- Fixed handling of `TokenAccountNotFoundError` within `splActionProvider` `getBalance`.
+- Fixed `wowActionProvider` exports, supported networks, and ensuring response parity with python.
 
 ## [0.2.1] - 2025-02-18